requirements.txt
flask==3.1.*
httpx==0.28.*
apscheduler==3.11.*
python-dotenv==1.1.*
gunicorn==23.*
rapidfuzz==3.*
